baru

+++ app/src/main/java/com/example/nfnt/fragment_week2/DetailActivity.java
@@ -2,6 +2,7 @@ package com.example.nfnt.fragment_week2;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.Window;
 
 public class DetailActivity extends AppCompatActivity {
 
@@ -9,10 +10,9 @@ public class DetailActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_detail);
-
         ResepDetailFragment frag = (ResepDetailFragment) getSupportFragmentManager().findFragmentById(R.id.detailFrag);
         Bundle b = getIntent().getExtras();
-        frag.setResep(b.getInt("id"));
+        frag.setResep(b.getLong("id"));
         //frag.setResep(1);
     }
 }
